Update README.md
General: Labs are done by individuals, not "us." 
I don't know what the tag "walkthrough-editor-spotlight" means. Can you just tag as code? 
Lines 12 - 16: Use bullets instead of numbers. The steps in the procedures, however, should be numbered. 
Line 12: Specify what "this app" is. 
Line 15: Is "the application" the same as "this app"? It's not clear. 
Line 24: Use numbers for procedure steps. 
Line 24: First tell the learner where to run this command. Is it in Cloud Shell? If so, tell the learner how to open Cloud Shell. See an example in go/lab-guide-guidelines.
Line 42: Deleted this line because it interrupts the flow of the procedure and doesn't offer new information. 
Line 50: This is not a user action so shouldn't be a step. Instead, indent it under the previous step as a result. 
Line 59: Does the learner do anything with this code, or is just showing where the Web Preview link is? If the latter, you can delete it (and adjust the previous line). 
Line 65: I deleted the "Next" sentence because the same information is in the next procedure. 
Line 85: This is not a user action so shouldn't be a step. Instead, indent it under the previous step as a result. That means that the procedure has only one step, so use a bullet for it instead of a number. 
Line 112: If the app returns "Hello world," why was "Hello Anthos" the expected result in line 101? Also, should "world" be capped or lowercase? These two instances are inconsistent. 
Line 123: This is not a user action so shouldn't be a step. Instead, indent it under the previous step as a result. 
Line 151: This is not a user action so shouldn't be a step. Instead, indent it under the previous step as a result. 
Line 159: This is not a user action either. Format it as a note. 
Line 194: What is the name of this button? Use this language: Click <b>button name</b> (icon image)...